Modif majeurs

+++ app/Tools/DevisTools.php
@@ -11,6 +11,7 @@ namespace App\Tools;
 
 use App\Demande;
 use App\Devi;
+use App\DevisEtat;
 use App\Etat;
 use App\Service;
 use App\User;
@@ -19,11 +20,13 @@ class DevisTools{
 
     public static function addDevis($request,User $u){
         $devis = new Devi();
+        $etat_devis = DevisEtatTools::getEtatByName('Créé');
         $demande = DemandeTools::getDemande($request['demande_id']);
         $interp = InterpreteurTools::getInterpreteur($request['interpreteur_id']);
         $devis->interpreteur()->associate($interp);
         $devis->demande()->associate($demande);
         $devis->user()->associate($u);
+        $devis->etat()->associate($etat_devis);
         $devis->save();
         $devis->total = ServiceTools::addServices($devis,$request);
         $devis->save();
@@ -34,6 +37,11 @@ class DevisTools{
 
     public static function getDevis($demande_id){
         $devis = Devi::where('demande_id',$demande_id)->get();
+        return $devis;
+    }
+
+    public static function getDevisById($id){
+        $devis = Devi::withTrashed()->where('id',$id)->get()->first();
         return $devis;
     }
 
@@ -60,7 +68,10 @@ class DevisTools{
     }
 
     public static function deleteDevis(User $u,$devis){
-        $devis->delete();
+        $facture = FactureTools::getFactureByDevis($devis);
+        if($facture == null) $devis->delete();
+        else return false;
+        return true;
     }
 
     public static function restoreDevis(User $u,$devis_id){
@@ -89,6 +100,20 @@ class DevisTools{
         $interpreteur = InterpreteurTools::getInterpreteur($devis->interpreteur_id);
         $demande = DemandeTools::getDemande($devis->demande_id);
         return DemandeTools::canBeRestored($demande) && !$interpreteur->trashed() && !$demande->trashed();
+    }
+
+    public static function validerDevis(User $u,$devis){
+        $etat = DevisEtatTools::getEtatByName('Commande');
+        $devis->etat()->associate($etat);
+        $devis->save();
+    }
+
+    public static function facturerDevis($devis){
+        $facture = FactureTools::addFacture($devis);
+        $etat = DevisEtatTools::getEtatByName('Validé');
+        $devis->etat()->associate($etat);
+        $devis->save();
+        return $facture;
     }
 
 }
